# Bound Check Elimination: BenchmarkAppendFastDecodeBase62 faster by 8%