```
benchmark                              old ns/op     new ns/op     delta
BenchmarkAppendFastDecodeBase62-12     143           131           -8.39%
```